requirements-test.txt
-r requirements.txt
coverage
flake8>=2.1.0
tox>=1.7.0
python-memcached
django-app-helper

